readched stable performance of algorithm
the implemented algorithm depends for its choice of elevator on the
total number of floors travelled by the each elevator until the person
reaches his destination floor. minimum is the best.
assumed rules:
1. elevator location is changed once the door is closed
2. the elevator can stop at intermidiate floor if there is a new request
came from it but without affecting the total number of floors travelled
by the person being served.